refactor : 프로필 공유 상태 수정 로직 분리 및 불필요 state 제거
