+++ src/1_Page/Profile/index.js
@@ -10,7 +10,6 @@ import useInfinityScroll from "./model/useInfinityScroll.js";
 
 import useGetTrackingImageList from "../../3_Entity/Tracking/useGetTrackingImageList.js";
 import { useParams } from "react-router-dom";
-import { useEffect, useRef } from "react";
 
 const Profile = () => {
   const { tabState, handleTabClick } = useTabs();
@@ -18,13 +17,8 @@ const Profile = () => {
 
   const { userIdx } = useParams(); // userIdx 추출
 
-  const containerRef = useRef(null);
-  const containerRef2 = useRef(null);
-
   const { paging, handleScroll, checkLessLength } = useInfinityScroll(
-    tabState.tabIndex,
-    containerRef,
-    containerRef2
+    tabState.tabIndex
   );
 
   const { trackingImageList, loading, hasMoreContent } =
@@ -37,22 +31,7 @@ const Profile = () => {
     handleSelectCancel,
     handleModifyTrack,
     handleDeleteTrack,
-  } = useManageTrackData(trackingImageList);
-
-  const prevLength = useRef(null);
-
-  useEffect(() => {
-    const track =
-      tabState.tabIndex === 0 ? shareTrackingImageData : saveTrackingImageData;
-
-    if (track.length === 0) return;
-
-    if (track.length !== prevLength.current) {
-      console.log(track.length);
-      checkLessLength(track.length);
-      prevLength.current = track.length; // 이전 값 저장
-    }
-  }, [shareTrackingImageData, saveTrackingImageData]);
+  } = useManageTrackData(trackingImageList, tabState.tabIndex, checkLessLength);
 
   return (
     <>
@@ -74,14 +53,12 @@ const Profile = () => {
               modifyMode={modifyMode}
               handleScroll={hasMoreContent.share ? handleScroll : null}
               handleAddModifyIdxList={handleAddModifyIdxList}
-              containerRef={containerRef}
             />
             <TrackingImageTab
               trackData={saveTrackingImageData}
               modifyMode={modifyMode}
               handleScroll={hasMoreContent.save ? handleScroll : null}
               handleAddModifyIdxList={handleAddModifyIdxList}
-              containerRef={containerRef2}
             />
           </STYLE.Slider>
         </STYLE.SliderWrapper>
