+++ src/1_Page/Profile/ui/TrackingImageTab/index.js
@@ -3,19 +3,13 @@ import STYLE from "./style";
 import TrackContainer from "./ui/TrackingImageContainer";
 
 const TrackingImageTab = (props) => {
-  const {
-    trackData,
-    modifyMode,
-    handleScroll,
-    handleAddModifyIdxList,
-    containerRef,
-  } = props;
+  const { trackData, modifyMode, handleScroll, handleAddModifyIdxList } = props;
   return (
     <>
       {trackData.length === 0 ? (
         <STYLE.EmptyMessage>게시물이 없습니다.</STYLE.EmptyMessage>
       ) : (
-        <STYLE.PostGrid ref={containerRef} onScroll={handleScroll}>
+        <STYLE.PostGrid onScroll={handleScroll}>
           {trackData.map((track) => {
             return (
               <TrackContainer
